Massive migration to a new engine
I call it the blerg-engine. It is my blerg-engine.
It's going to be a work in progress perpetually for as long as I
need it. Adding functionality as it becomes important (RSS, special
features, etc)

For the time being it kinda gets the job done. It's simple and that
is all I need it to do. K.I.S.S. and such.